major changes
Combined Fits:
added function to overlay a chosen histogram from a list of files
working.

ParticleDecayTest2:

removed the need to have a spectrum histogram. my cluster overlay was drawing from a SMEARED and weighted spectrum, which I don't think is correct to use.
I am drawing from a TF1.
That TF1 is generated in an outside function depending on the weight method int supplied to the function