lots of work
added clusterover lay test to aiofit. needs review
made histograms for all weights in a single file. now the only difference in files is the smearing percent.
maybe I can add that eventually too?
it seems to be working for now, but I have to change it so that the spectrum function is also a vector. for now I still select it at the start.
I am also looping and making 1 file per weight right now.